Fixed loginPage and analysisToolsPage.  Add validations into programs.

+++ lib/src/models/new_parts_model.dart
@@ -36,7 +36,7 @@ class NewPartModel {
   int programsId;
   int typesSizesId;
   String name;
-  int plannedLines;
+  double plannedLines;
   int numberMethodsPlanned;
   int currentLines;
   int numberMethodsCurrent;
